requirements.txt
pyyaml
numpy
torchvision
torch
matplotlib
multiprocessing